% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompRiskRE-package.R
\docType{package}
\name{CompRiskRE-package}
\alias{CompRiskRE}
\alias{CompRiskRE-package}
\title{CompRiskRE: Relative Entropy Integration for Competing Risks Models}
\description{
The CompRiskRE package provides functions for discrete-time competing risks modeling
with Relative Entropy (RE) integration. It supports both composite failure time (CFT) 
prior models and cause-specific prior models, and includes simulation, model fitting,
and performance evaluation utilities.
}
\section{Installation}{

You can install the package from GitHub:
\preformatted{
  devtools::install_github("UM-KevinHe/CompRiskRE")
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UM-KevinHe/CompRiskRE}
  \item \url{https://umkevinhe.github.io/CompRiskRE}
}

}
\author{
\strong{Maintainer}: Yubo Shao \email{ybshao@umich.edu}

Authors:
\itemize{
  \item Qinmengge Li \email{liqinmg@umich.edu}
}

}
\keyword{internal}
