% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompRiskRE_XGBoost_FT.R
\name{CompRiskRE_XGBoost_FT}
\alias{CompRiskRE_XGBoost_FT}
\title{Competing Risks with Relative Entropy Integration using XGBoost and Composite Failure Time Prior}
\usage{
CompRiskRE_XGBoost_FT(
  prior_model,
  train_data,
  test_data,
  eta,
  xgb_params = list(objective = "multi:softprob", eval_metric = "mlogloss", num_class =
    3, max.depth = 4),
  nrounds = 100,
  maxiter = 1,
  eps = 1e-06
)
}
\arguments{
\item{prior_model}{A fitted prior multinomial model.}

\item{train_data}{A \code{data.frame} containing training data.}

\item{test_data}{A \code{data.frame} containing test data.}

\item{eta}{Numeric vector of RE regularization parameters.}

\item{xgb_params}{A list of XGBoost parameters..}

\item{nrounds}{Integer, number of boosting rounds for XGBoost (default: 100).}

\item{maxiter}{Integer, maximum number of RE updates in \code{priorFTKL_XGBoost} (default: 1).}

\item{eps}{Numeric tolerance for convergence in \code{priorFTKL_XGBoost} (default: 1e-6).}
}
\value{
A list with two components:
\describe{
  \item{models}{A named list of fitted XGBoost models for each supplied \code{eta}.}
  \item{PD}{A numeric vector of predictive deviance values on the test data.}
  \item{eta}{The vector of \code{eta} values.}
}
}
\description{
This function implements the Relative Entropy (RE) framework for discrete-time competing risks models
where the prior model is specified in a composite failure time formulation and the local model is fit
using XGBoost.
}
