% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_eta}
\alias{generate_eta}
\title{Generate a Sequence of Tuning Parameters (eta)}
\usage{
generate_eta(method = "exponential", n = 10, max_eta = 5)
}
\arguments{
\item{method}{Character string specifying the method to generate `eta`.
Options are `"linear"` for a linearly spaced sequence, or `"exponential"`
for an exponentially spaced sequence scaled to `max_eta`. Default is `"exponential"`.}

\item{n}{Integer, the number of `eta` values to generate. Default is 10.}

\item{max_eta}{Numeric, the maximum value of `eta` in the sequence. Default is 5.}
}
\value{
Numeric vector of length `n` containing the generated `eta` values.
}
\description{
Generate a Sequence of Tuning Parameters (eta)
}
\examples{
# Generate 10 exponentially spaced eta values up to 5
generate_eta(method = "exponential", n = 10, max_eta = 5)

# Generate 5 linearly spaced eta values up to 3
generate_eta(method = "linear", n = 5, max_eta = 3)

}
