% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_n}
\alias{prob_n}
\title{Simulate event times from a discrete competing risks model}
\usage{
prob_n(x, betat, betav, N, p, Tmax, nCause, unif = TRUE)
}
\arguments{
\item{x}{Covariate matrix (\eqn{N \times p}).}

\item{betat}{Time-related parameters.}

\item{betav}{Covariate-related parameters.}

\item{N}{Number of individuals.}

\item{p}{Number of covariates.}

\item{Tmax}{Maximum number of discrete time points.}

\item{nCause}{Number of competing risks.}

\item{unif}{Logical; if \code{TRUE}, the last column is sampled uniformly, otherwise from a logistic model.}
}
\value{
A numeric matrix of dimension \eqn{N \times (nCause+1)} where each entry
  is the simulated event time (or censoring time) for a given individual and risk.
}
\description{
Generates event times for multiple competing risks under a discrete-time
hazard specification.
}
